Restore v1alpha1 DevWorkspace CRD
The v1alpha1 apiVersion of the DevWorkspace (and DevWorkspaceTemplate)
CRD was removed prematurely as there are existing users that depend on
it.

This commit restores the v1alpha1 API as removed in PR #639 in order to
allow dependents of the DevWorkspace API to update to later versions of
the v1alpha2 API without breaking v1alpha1 compatibility.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>