Allow duplicate endpoint targetPorts if they are in same container
Allow multiple endpoints to have the same target point if and only if
they are attached to the same container. Otherwise, target ports must be
unique unless dedicatedPod is set to true.

Allowing multiple endpoints to use the same port is required for some
applications which require two routes to the same endpoint, in order to
create isolation between components within the browser.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>