fix (utils) : [CRW-8603] Use slice of string instead of apimachinery Set to merge events
We were previously using apimachinery Set that is internally based on Map, an unordered data structure.
Due to this we were losing the original insertion order defined in the devfile while merging it, List()
call returns the elements in non-deterministic order, as it internally ranges over the map.

Use string slice to preserve insertion order of all event types.

Signed-off-by: Rohan Kumar <rohaan@redhat.com>